remove permissions. Reorganize apis and tests
