+++ src/sentry/api/urls.py
@@ -293,11 +293,17 @@ from sentry.integrations.api.endpoints.organization_repository_platforms import 
 from sentry.integrations.api.endpoints.organization_repository_settings import (
     OrganizationRepositorySettingsEndpoint,
 )
-from sentry.investigations.endpoints.organization_investigations import (
-    OrganizationInvestigationDetailsEndpoint,
-    OrganizationInvestigationDuplicateEndpoint,
-    OrganizationInvestigationFavoriteEndpoint,
-    OrganizationInvestigationsEndpoint,
+from sentry.investigations.endpoints.organization_investigations_details import (
+    OrganizationInvestigationsDetailsEndpoint,
+)
+from sentry.investigations.endpoints.organization_investigations_duplicate import (
+    OrganizationInvestigationsDuplicateEndpoint,
+)
+from sentry.investigations.endpoints.organization_investigations_favorite import (
+    OrganizationInvestigationsFavoriteEndpoint,
+)
+from sentry.investigations.endpoints.organization_investigations_index import (
+    OrganizationInvestigationsIndexEndpoint,
 )
 from sentry.issues.endpoints import (
     ActionableItemsEndpoint,
@@ -2370,22 +2376,22 @@ ORGANIZATION_URLS: list[URLPattern | URLResolver] = [
     ),
     re_path(
         r"^(?P<organization_id_or_slug>[^/]+)/investigations/$",
-        OrganizationInvestigationsEndpoint.as_view(),
+        OrganizationInvestigationsIndexEndpoint.as_view(),
         name="sentry-api-0-organization-investigations",
     ),
     re_path(
         r"^(?P<organization_id_or_slug>[^/]+)/investigations/(?P<investigation_id>[^/]+)/$",
-        OrganizationInvestigationDetailsEndpoint.as_view(),
+        OrganizationInvestigationsDetailsEndpoint.as_view(),
         name="sentry-api-0-organization-investigation-details",
     ),
     re_path(
         r"^(?P<organization_id_or_slug>[^/]+)/investigations/(?P<investigation_id>[^/]+)/favorite/$",
-        OrganizationInvestigationFavoriteEndpoint.as_view(),
+        OrganizationInvestigationsFavoriteEndpoint.as_view(),
         name="sentry-api-0-organization-investigation-favorite",
     ),
     re_path(
         r"^(?P<organization_id_or_slug>[^/]+)/investigations/(?P<investigation_id>[^/]+)/duplicate/$",
-        OrganizationInvestigationDuplicateEndpoint.as_view(),
+        OrganizationInvestigationsDuplicateEndpoint.as_view(),
         name="sentry-api-0-organization-investigation-duplicate",
     ),
     re_path(
