+++ src/sentry/investigations/endpoints/organization_investigations_details.py
@@ -1,0 +1,108 @@
+from __future__ import annotations
+
+from drf_spectacular.utils import extend_schema
+from rest_framework import status
+from rest_framework.request import Request
+from rest_framework.response import Response
+
+from sentry.api.api_publish_status import ApiPublishStatus
+from sentry.api.base import cell_silo_endpoint
+from sentry.api.serializers import serialize
+from sentry.investigations.endpoints.bases import (
+    OrganizationInvestigationEndpoint,
+    accessible_project_ids,
+    service_error,
+)
+from sentry.investigations.endpoints.serializers import InvestigationDetailsSerializer
+from sentry.investigations.endpoints.validators import (
+    InvestigationDeleteValidator,
+    InvestigationUpdateValidator,
+)
+from sentry.investigations.models import Investigation, InvestigationStatus
+from sentry.investigations.services import archive_investigation, update_investigation
+from sentry.models.organization import Organization
+
+
+@extend_schema(tags=["Investigations"])
+@cell_silo_endpoint
+class OrganizationInvestigationsDetailsEndpoint(OrganizationInvestigationEndpoint):
+    publish_status = {
+        "GET": ApiPublishStatus.PRIVATE,
+        "PUT": ApiPublishStatus.PRIVATE,
+        "DELETE": ApiPublishStatus.PRIVATE,
+    }
+
+    def get(
+        self, request: Request, organization: Organization, investigation: Investigation
+    ) -> Response:
+        return Response(
+            serialize(
+                investigation,
+                request.user,
+                InvestigationDetailsSerializer(
+                    accessible_project_ids=accessible_project_ids(self, request, organization)
+                ),
+            )
+        )
+
+    def put(
+        self, request: Request, organization: Organization, investigation: Investigation
+    ) -> Response:
+        validator = InvestigationUpdateValidator(data=request.data)
+        if not validator.is_valid():
+            return Response(validator.errors, status=status.HTTP_400_BAD_REQUEST)
+        values = dict(validator.validated_data)
+        expected_version = values.pop("investigation_version")
+        requested_project_ids = values.pop("project_ids", None)
+        project_ids = accessible_project_ids(self, request, organization)
+        if requested_project_ids is not None and not set(requested_project_ids).issubset(
+            project_ids
+        ):
+            return Response(
+                {"detail": "One or more projects are inaccessible."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+        if investigation.status == InvestigationStatus.ARCHIVED and values != {
+            "status": InvestigationStatus.ACTIVE
+        }:
+            return Response(
+                {"detail": "Archived investigations are read-only."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+        try:
+            updated = update_investigation(
+                investigation=investigation,
+                expected_version=expected_version,
+                fields=values,
+                project_ids=requested_project_ids,
+            )
+        except Exception as error:
+            response = service_error(error)
+            if response is not None:
+                return response
+            raise
+        return Response(
+            serialize(
+                updated,
+                request.user,
+                InvestigationDetailsSerializer(accessible_project_ids=project_ids),
+            )
+        )
+
+    def delete(
+        self, request: Request, organization: Organization, investigation: Investigation
+    ) -> Response:
+        validator = InvestigationDeleteValidator(data=request.data)
+        if not validator.is_valid():
+            return Response(validator.errors, status=status.HTTP_400_BAD_REQUEST)
+        try:
+            archive_investigation(
+                investigation=investigation,
+                expected_version=validator.validated_data["investigation_version"],
+            )
+        except Exception as error:
+            response = service_error(error)
+            if response is not None:
+                return response
+            raise
+        return Response(status=status.HTTP_204_NO_CONTENT)
