+++ src/sentry/investigations/models/__init__.py
@@ -4,4 +4,3 @@ from .cell import *  # NOQA
 from .execution import *  # NOQA
 from .investigation import *  # NOQA
 from .parameter import *  # NOQA
-from .permissions import *  # NOQA
