+++ tests/sentry/investigations/endpoints/test_organization_investigations_index.py
@@ -1,0 +1,188 @@
+from __future__ import annotations
+
+from unittest import mock
+
+from django.urls import reverse
+
+from sentry.investigations.models import (
+    Investigation,
+    InvestigationSourceType,
+    InvestigationStatus,
+)
+from sentry.investigations.templates.types import (
+    InvestigationTemplateSpec,
+    TemplateBlockSpec,
+)
+from sentry.testutils.cases import APITestCase
+from sentry.testutils.helpers.features import with_feature
+
+FEATURE = "organizations:investigations"
+
+
+@with_feature(FEATURE)
+class OrganizationInvestigationsIndexTest(APITestCase):
+    def setUp(self) -> None:
+        super().setUp()
+        self.login_as(self.user)
+        self.collection_url = reverse(
+            "sentry-api-0-organization-investigations",
+            kwargs={"organization_id_or_slug": self.organization.slug},
+        )
+
+    def test_create_manual_and_list(self) -> None:
+        response = self.client.post(
+            self.collection_url,
+            data={
+                "title": "Checkout follow-up",
+                "projectIds": [self.project.id],
+                "filters": {"environment": ["production"]},
+            },
+            format="json",
+        )
+        assert response.status_code == 201, response.data
+        assert response.data["title"] == "Checkout follow-up"
+        assert response.data["projectIds"] == [self.project.id]
+        assert response.data["blocks"] == []
+
+        response = self.client.get(self.collection_url)
+        assert response.status_code == 200
+        assert [item["title"] for item in response.data] == ["Checkout follow-up"]
+        assert response.data[0]["blockCount"] == 0
+        assert response.data[0]["isFavorited"] is False
+
+    def test_manual_creation_rejects_inaccessible_project(self) -> None:
+        other_organization = self.create_organization()
+        other_project = self.create_project(organization=other_organization)
+        response = self.client.post(
+            self.collection_url,
+            data={"title": "No access", "projectIds": [other_project.id]},
+            format="json",
+        )
+        assert response.status_code == 400
+        assert not Investigation.objects.filter(title="No access").exists()
+
+    def test_template_validation_is_strict_and_atomic(self) -> None:
+        before = Investigation.objects.count()
+        response = self.client.post(
+            self.collection_url,
+            data={
+                "templateKey": "breached_metric",
+                "templateVersion": 1,
+                "sourceRef": {"groupId": "1", "openPeriodId": "1"},
+                "parameters": {"unexpected": True},
+            },
+            format="json",
+        )
+        assert response.status_code == 400
+        assert "parameters" in response.data
+        assert Investigation.objects.count() == before
+
+    def test_unknown_template_version_is_atomic(self) -> None:
+        before = Investigation.objects.count()
+        response = self.client.post(
+            self.collection_url,
+            data={
+                "templateKey": "breached_metric",
+                "templateVersion": 999,
+                "sourceRef": {"groupId": "1", "openPeriodId": "1"},
+                "parameters": {},
+            },
+            format="json",
+        )
+        assert response.status_code == 400
+        assert Investigation.objects.count() == before
+
+    def test_cyclic_template_rolls_back_before_source_resolution(self) -> None:
+        template = InvestigationTemplateSpec(
+            key="cyclic",
+            version=1,
+            source_type=InvestigationSourceType.BREACHED_METRIC,
+            parameters=(),
+            blocks=(
+                TemplateBlockSpec(key="one", kind="text", title="One", dependencies=("two",)),
+                TemplateBlockSpec(key="two", kind="text", title="Two", dependencies=("one",)),
+            ),
+        )
+        before = Investigation.objects.count()
+        with mock.patch(
+            "sentry.investigations.services.investigations.get_investigation_template",
+            return_value=template,
+        ):
+            response = self.client.post(
+                self.collection_url,
+                data={
+                    "templateKey": "cyclic",
+                    "templateVersion": 1,
+                    "sourceRef": {"groupId": "1"},
+                    "parameters": {},
+                },
+                format="json",
+            )
+        assert response.status_code == 400
+        assert Investigation.objects.count() == before
+
+    def test_template_rejects_wrong_issue_category(self) -> None:
+        group = self.create_group(project=self.project)
+        response = self.client.post(
+            self.collection_url,
+            data={
+                "templateKey": "breached_metric",
+                "templateVersion": 1,
+                "sourceRef": {"groupId": str(group.id), "openPeriodId": "1"},
+                "parameters": {},
+            },
+            format="json",
+        )
+        assert response.status_code == 404
+
+    def test_source_lineage_lists_latest_and_keeps_historical_detail(self) -> None:
+        lineage = {
+            "organization": self.organization,
+            "created_by": self.user,
+            "source_type": InvestigationSourceType.ISSUE,
+            "source_key": "issue:123",
+            "source_ref": {"groupId": "123"},
+        }
+        first = self.create_investigation(
+            title="First revision", source_revision=1, status=InvestigationStatus.ACTIVE, **lineage
+        )
+        second = self.create_investigation(
+            title="Second revision", source_revision=2, status=InvestigationStatus.ACTIVE, **lineage
+        )
+
+        response = self.client.get(self.collection_url)
+        assert [item["id"] for item in response.data] == [str(second.id)]
+
+        first_url = reverse(
+            "sentry-api-0-organization-investigation-details",
+            kwargs={
+                "organization_id_or_slug": self.organization.slug,
+                "investigation_id": first.id,
+            },
+        )
+        assert self.client.get(first_url).status_code == 200
+
+        second_url = reverse(
+            "sentry-api-0-organization-investigation-details",
+            kwargs={
+                "organization_id_or_slug": self.organization.slug,
+                "investigation_id": second.id,
+            },
+        )
+        assert (
+            self.client.delete(
+                second_url,
+                data={"investigationVersion": second.version},
+                format="json",
+            ).status_code
+            == 204
+        )
+        assert not Investigation.objects.filter(
+            source_key="issue:123", status=InvestigationStatus.ACTIVE
+        ).exists()
+
+        third = self.create_investigation(
+            title="Third revision", source_revision=3, status=InvestigationStatus.ACTIVE, **lineage
+        )
+        response = self.client.get(self.collection_url)
+        assert [item["id"] for item in response.data] == [str(third.id)]
